pyproject.toml
[tool.poetry]
name = "craft-be"
version = "0.1.0"
description = ""
authors = ["srk0102 <ramakrishnasiva128@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
user-agents = "^2.2.0"
requests = "^2.32.3"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
greenlet = "^3.1.1"
celery = {extras = ["redis"], version = "^5.4.0"}
pika = "^1.3.2"
motor = "^3.6.0"
certifi = "^2024.8.30"
email-validator = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
